# packages/types/openapi.json
{
  "components": {
    "schemas": {
      "ActionBrief": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "status": {
            "$ref": "#/components/schemas/ActionStatus"
          },
          "type": {
            "$ref": "#/components/schemas/ActionType"
          }
        },
        "required": [
          "id",
          "type",
          "status",
          "created_at",
          "attempts",
          "last_error"
        ],
        "title": "ActionBrief",
        "type": "object"
      },
...
      "ActionStatus": {
        "enum": [
          "PENDING",
          "PROCESSING",
          "SUCCESS",
          "FAILED",
          "NEEDS_ATTENTION"
        ],
        "title": "ActionStatus",
        "type": "string"
      },
      "ActionType": {
        "enum": [
          "SEND_MESSAGE",
          "MARK_ORDER_SHIPPED",
          "PAIR_SESSION",
          "VERIFY_SESSION"
        ],
        "title": "ActionType",
        "type": "string"
      },
...
      "ConversationRef": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_message_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message At"
          },
          "unread": {
            "title": "Unread",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "unread",
          "last_message_at"
        ],
        "title": "ConversationRef",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItemOut": {
        "properties": {
          "card_name": {
            "title": "Card Name",
            "type": "string"
          },
          "condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition"
          },
          "expansion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expansion"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "total_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Price"
          },
          "unit_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          }
        },
        "required": [
          "id",
          "card_name",
          "expansion",
          "language",
          "condition",
          "quantity",
          "unit_price",
          "total_price"
        ],
        "title": "ItemOut",
        "type": "object"
      },
...
      "NormalizedOrder": {
        "additionalProperties": false,
        "description": "Full order detail.",
        "properties": {
          "buyer_name": {
            "title": "Buyer Name",
            "type": "string"
          },
          "cardmarket_id": {
            "minLength": 1,
            "title": "Cardmarket Id",
            "type": "string"
          },
          "currency": {
            "default": "EUR",
            "title": "Currency",
            "type": "string"
          },
          "extra": {
            "additionalProperties": true,
            "title": "Extra",
            "type": "object"
          },
          "item_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/NormalizedItem"
            },
            "title": "Items",
            "type": "array"
          },
          "order_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Date"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "payment_status": {
            "$ref": "#/components/schemas/PaymentStatus",
            "default": "UNKNOWN"
          },
          "shipped_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipped At"
          },
          "shipping_status": {
            "$ref": "#/components/schemas/ShippingStatus",
            "default": "UNKNOWN"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          },
          "tracking_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Number"
          }
        },
        "required": [
          "cardmarket_id",
          "buyer_name",
          "status"
        ],
        "title": "NormalizedOrder",
        "type": "object"
      },
...
      "OrderDetail": {
        "properties": {
          "buyer_name": {
            "title": "Buyer Name",
            "type": "string"
          },
          "cardmarket_id": {
            "title": "Cardmarket Id",
            "type": "string"
          },
          "conversations": {
            "items": {
              "$ref": "#/components/schemas/ConversationRef"
            },
            "title": "Conversations",
            "type": "array"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "detail_fetched_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail Fetched At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_new": {
            "title": "Is New",
            "type": "boolean"
          },
          "item_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "order_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Date"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "payment_status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "pending_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "raw_source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Source Reference"
          },
          "shipped_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipped At"
          },
          "shipping_status": {
            "$ref": "#/components/schemas/ShippingStatus"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "total_amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          },
          "tracking_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Number"
          }
        },
        "required": [
          "id",
          "cardmarket_id",
          "buyer_name",
          "status",
          "payment_status",
          "shipping_status",
          "total_amount",
          "currency",
          "item_count",
          "order_date",
          "paid_at",
          "shipped_at",
          "is_new",
          "last_seen_at",
          "tracking_number",
          "raw_source_reference",
          "detail_fetched_at",
          "items",
          "conversations"
        ],
        "title": "OrderDetail",
        "type": "object"
      },
      "OrderFilter": {
        "enum": [
          "all",
          "new",
          "paid",
          "unpaid",
          "to_ship",
          "shipped",
          "completed",
          "cancelled"
        ],
        "title": "OrderFilter",
        "type": "string"
      },
      "OrderListItem": {
        "properties": {
          "buyer_name": {
            "title": "Buyer Name",
            "type": "string"
          },
          "cardmarket_id": {
            "title": "Cardmarket Id",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_new": {
            "title": "Is New",
            "type": "boolean"
          },
          "item_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Count"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "order_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Date"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "payment_status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "shipped_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipped At"
          },
          "shipping_status": {
            "$ref": "#/components/schemas/ShippingStatus"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "total_amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          }
        },
        "required": [
          "id",
          "cardmarket_id",
          "buyer_name",
          "status",
          "payment_status",
          "shipping_status",
          "total_amount",
          "currency",
          "item_count",
          "order_date",
          "paid_at",
          "shipped_at",
          "is_new",
          "last_seen_at"
        ],
        "title": "OrderListItem",
        "type": "object"
      },
...
      "OrderSort": {
        "enum": [
          "date_desc",
          "date_asc",
          "total_desc",
          "total_asc",
          "buyer"
        ],
        "title": "OrderSort",
        "type": "string"
      },
      "OrderStatus": {
        "description": "Internal order lifecycle, independent from Cardmarket wording.",
        "enum": [
          "UNPAID",
          "PAID",
          "SHIPPED",
          "COMPLETED",
          "CANCELLED",
          "UNKNOWN"
        ],
        "title": "OrderStatus",
        "type": "string"
      },
      "OrdersBatch": {
        "properties": {
          "details": {
            "items": {
              "$ref": "#/components/schemas/NormalizedOrder"
            },
            "title": "Details",
            "type": "array"
          },
          "summaries": {
            "items": {
              "$ref": "#/components/schemas/NormalizedOrderSummary"
            },
            "title": "Summaries",
            "type": "array"
          }
        },
        "title": "OrdersBatch",
        "type": "object"
      },
...
      "Page_OrderListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[OrderListItem]",
        "type": "object"
      },
...
      "PaymentStatus": {
        "enum": [
          "UNPAID",
          "PAID",
          "REFUNDED",
          "UNKNOWN"
        ],
        "title": "PaymentStatus",
        "type": "string"
      },
...
      "ShipOrderRequest": {
        "properties": {
          "client_key": {
            "maxLength": 64,
            "minLength": 8,
            "title": "Client Key",
            "type": "string"
          },
          "tracking_number": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Number"
          }
        },
        "required": [
          "client_key"
        ],
        "title": "ShipOrderRequest",
        "type": "object"
      },
      "ShippingStatus": {
        "enum": [
          "NOT_SHIPPED",
          "SHIPPED",
          "DELIVERED",
          "UNKNOWN"
        ],
        "title": "ShippingStatus",
        "type": "string"
      },
...
      "SyncBatchResult": {
        "properties": {
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "records_changed": {
            "title": "Records Changed",
            "type": "integer"
          },
          "records_found": {
            "title": "Records Found",
            "type": "integer"
          }
        },
        "required": [
          "records_found",
          "records_changed",
          "events"
        ],
        "title": "SyncBatchResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cardmarket Companion API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/orders": {
      "get": {
        "operationId": "list_orders_api_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OrderFilter",
              "default": "all"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OrderSort",
              "default": "date_desc"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_OrderListItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Orders",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{order_id}": {
      "get": {
        "operationId": "get_order_api_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{order_id}/acknowledge": {
      "post": {
        "operationId": "acknowledge_order_api_orders__order_id__acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acknowledge Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{order_id}/ship": {
      "post": {
        "description": "Queue 'mark as shipped' on Cardmarket. Local status changes only once verified.",
        "operationId": "ship_order_api_orders__order_id__ship_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionBrief"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ship Order",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/internal/agent/sync/{run_id}/orders": {
      "post": {
        "operationId": "sync_orders_internal_agent_sync__run_id__orders_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdersBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncBatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Orders",
        "tags": [
          "internal-agent"
        ]
      }
    }
  }
}